Promote Transaction to a top-level class and make membership explicit
