+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
 #### Version 2.3.0 (unreleased)
-* **Added `Runway#transaction` to scope reads and writes to a single ACID transaction.** `transaction()` returns a `Transaction`: a `DatabaseInterface` view whose reads observe the transaction's isolated snapshot, including its own uncommitted writes, and join its conflict footprint, so a commit fails instead of persisting a decision that was made on data a concurrent writer changed. ([#165](https://github.com/cinchapi/runway/pull/165))
-    * Every `Record` loaded through the view, including the linked records reachable from it, is bound to the transaction: `save()` stages within it and the writes become durable only when `commit()` succeeds. New records join through `Transaction#save`. Save and delete notifications dispatch only after a successful commit. An abort discards every staged write; a record's in-memory edits remain, the same as after a failed save.
-    * `Transaction` is `AutoCloseable` (`close()` aborts whatever was not committed) and is confined to the thread that starts it. After it ends, the view forwards reads and saves to the enclosing `Runway`, so bound records unwind to the database scope; only another `commit()` is refused. `runway.run(work)` executes work within a transaction and commits after it completes: the work reads through the provided `DatabaseInterface`, every `record.save()` during the work joins the transaction, new records included, and conflicts retry within the bounds of the governing `AtomicRetryPolicy`.
-    * An `Audience` loaded through the view routes the operations it performs through the transaction, so access-controlled reads and writes stay within the snapshot and join the conflict footprint.
+* **Added `Runway#stage` to scope reads and writes to a single ACID transaction.** `stage()` (or its alias, `startTransaction()`) returns a `Transaction`: a `DatabaseInterface` view whose reads observe the transaction's isolated snapshot, including its own uncommitted writes, and join its conflict footprint, so a commit fails instead of persisting a decision that was made on data a concurrent writer changed. ([#165](https://github.com/cinchapi/runway/pull/165))
+    * Every `Record` loaded through the view, including the linked records reachable from it, is bound to the transaction: `save()` stages within it and the writes become durable only when `commit()` succeeds. New records join through `Transaction#save`, or by being created with `Transaction#create`, which binds them to the transaction so a direct `save()` stages within it. Save and delete notifications dispatch only after a successful commit. An abort discards every staged write; a record's in-memory edits remain, the same as after a failed save.
+    * `Transaction` is `AutoCloseable` (`close()` aborts whatever was not committed) and is confined to the thread that starts it. After it ends, the view forwards reads and saves to the enclosing `Runway`, so bound records unwind to the database scope; only another `commit()` is refused. `runway.run(work)` and `runway.call(work)` execute work within a transaction and commit after it completes: the work receives the open `Transaction` (`run` for work with no result, `call` for work that returns one), records loaded through it or created by it save within it, and conflicts retry within the bounds of the governing `AtomicRetryPolicy`. A record bound elsewhere saves against its own binding, outside of the transaction.
+    * An `Audience` loaded through the view routes the operations it performs through the transaction, so access-controlled reads and writes stay within the snapshot and join the conflict footprint. A record created through `Audience#create` is bound to the same context the audience operates against, so an audience within a transaction creates records whose saves stage within it.
     * Single-key atomic operations (`getAndUpdate`, `updateAndGet`, `exchange`) require a direct `Runway` binding and are refused on a transaction-bound `Record`, because the transaction's commit is the unit of atomicity. A `DeferredReference` that is first accessed within a transaction resolves within it: the lazy load joins the snapshot and the conflict footprint, and the loaded record binds to the transaction.
     * Internally, a `Record` now holds exactly two collaborators: a binding that reads and saves route through, and a `ConcourseProvider` that supplies connections for the record's own Concourse operations. `Record` never starts or commits a transaction.
