+++ src/main/java/com/cinchapi/runway/Runway.java
@@ -490,7 +490,7 @@ public final class Runway implements
      * This functionality is supported in Concourse 1.0.0+
      * </p>
      */
-    private final boolean supportsBulkCommands;
+    /* package */ final boolean supportsBulkCommands;
 
     /**
      * A flag that indicates if the connected server supports the {@code *}
@@ -552,14 +552,6 @@ public final class Runway implements
      * {@link #unreserve()}.
      */
     private final ThreadLocal<Map<Reservation, Object>> reservations = new ThreadLocal<>();
-
-    /**
-     * The open {@link Transaction} that {@link #run(Consumer) run} is executing
-     * on the current thread, if any. While one is present, every
-     * {@link #save(boolean, Record...) save} on the thread routes into it, so
-     * records saved during the work, including new ones, join the transaction.
-     */
-    private final ThreadLocal<Transaction> ambient = new ThreadLocal<>();
 
     /**
      * The executor for running isolated selections concurrently.
@@ -1031,7 +1023,7 @@ public final class Runway implements
      * @param saver the {@link Saver} for the active staged transaction
      * @param context the {@link SaveContext} for the save
      */
-    private void stageDeletions(Saver saver, SaveContext context) {
+    /* package */ void stageDeletions(Saver saver, SaveContext context) {
         Set<Long> deletions = context.deletions();
         if(!deletions.isEmpty()) {
             // NOTE: A deletion is final within a save. A record staged before
@@ -1062,7 +1054,7 @@ public final class Runway implements
      *
      * @param context the {@link SaveContext} whose staged save committed
      */
-    private void dispatchSaveOutcomes(SaveContext context) {
+    /* package */ void dispatchSaveOutcomes(SaveContext context) {
         Set<Long> deletions = context.deletions();
         context.forEach((record, outcome) -> {
             if(outcome == SaveContext.Outcome.DELETED) {
@@ -1165,10 +1157,6 @@ public final class Runway implements
      *             and any {@link Record} has been externally modified
      */
     public boolean save(boolean preventStaleWrites, Record... records) {
-        Transaction transaction = ambient.get();
-        if(transaction != null) {
-            return transaction.save(preventStaleWrites, records);
-        }
         Concourse concourse = connections.request();
         Record current = null;
         try {
@@ -2014,7 +2002,7 @@ public final class Runway implements
      * @param <T> the {@link Record} type
      */
     @SuppressWarnings({ "rawtypes" })
-    private <T extends Record> void $selectFromDatabase(Reader reader,
+    /* package */ <T extends Record> void $selectFromDatabase(Reader reader,
             DatabaseSelection<T> selection, Transaction transaction) {
         Pending<? extends SelectResult<?>> pending;
         if(selection instanceof CountSelection) {
@@ -2951,7 +2939,8 @@ public final class Runway implements
             // per attempt would churn the pool.
             int attempts = 0;
             for (;;) {
-                Transaction transaction = new Transaction(concourse, false);
+                Transaction transaction = new Transaction(this, concourse,
+                        false);
                 try {
                     T record;
                     if(order != null) {
@@ -3797,6 +3786,13 @@ public final class Runway implements
      */
     private class NullTransaction extends Transaction {
 
+        /**
+         * Construct a new instance.
+         */
+        NullTransaction() {
+            super(Runway.this);
+        }
+
         @Override
         public Selections select(Selection<?>... options) {
             return Runway.this.select(options);
@@ -3914,19 +3910,33 @@ public final class Runway implements
      *
      * @return an open {@link Transaction}
      */
-    public Transaction transaction() {
-        return new Transaction(connections.request(), true);
+    public Transaction stage() {
+        return new Transaction(this, connections.request(), true);
+    }
+
+    /**
+     * Start a {@link Transaction} that scopes reads and writes to a single ACID
+     * transaction.
+     * <p>
+     * This method is an alias for {@link #stage()}.
+     * </p>
+     *
+     * @return an open {@link Transaction}
+     */
+    public Transaction startTransaction() {
+        return stage();
     }
 
     /**
      * Run {@code work} within a {@link Transaction} and commit it after the
      * work completes.
      * <p>
-     * The work reads through the provided {@link DatabaseInterface}, so it
-     * observes the transaction's isolated snapshot. Every {@link Record#save()
-     * save} that happens during the work, including a save of a new
-     * {@link Record}, joins the transaction, and everything becomes durable
-     * together when the commit succeeds.
+     * The work receives the open {@link Transaction}: reads through it observe
+     * the transaction's isolated snapshot, and a {@link Record} loaded through
+     * it or {@link Transaction#create(Class, Object...) created} by it saves
+     * within it, so everything becomes durable together when the commit
+     * succeeds. A {@link Record} bound elsewhere saves against its own binding,
+     * outside of the transaction.
      * </p>
      * <p>
      * If the commit fails because of a conflict, then the transaction is
@@ -3938,10 +3948,11 @@ public final class Runway implements
      * </p>
      *
      * @param work the work to run
+     * @return the result of {@code work}
      * @throws RetryExhaustedException if the transaction cannot commit within
      *             the bounds of the governing {@link AtomicRetryPolicy}
      */
-    public void run(Consumer<DatabaseInterface> work) {
+    public <T> T call(Function<Transaction, T> work) {
         AtomicRetryPolicy policy = properties().atomicRetryPolicy();
         Concourse concourse = connections.request();
         try {
@@ -3950,25 +3961,18 @@ public final class Runway implements
             // per attempt would churn the pool.
             int attempts = 0;
             for (;;) {
-                Transaction transaction = new Transaction(concourse, false);
-                Transaction previous = ambient.get();
-                ambient.set(transaction);
+                Transaction transaction = new Transaction(this, concourse,
+                        false);
                 try {
-                    work.accept(transaction);
+                    T result = work.apply(transaction);
                     if(transaction.commit()) {
-                        return;
+                        return result;
                     }
                 }
                 catch (TransactionException e) {
                     transaction.abort();
                 }
                 finally {
-                    if(previous != null) {
-                        ambient.set(previous);
-                    }
-                    else {
-                        ambient.remove();
-                    }
                     transaction.close();
                 }
                 if(++attempts > policy.limit()) {
@@ -3985,388 +3989,22 @@ public final class Runway implements
     }
 
     /**
-     * A {@link DatabaseInterface} that scopes every operation to a single ACID
-     * transaction.
+     * Run {@code work} within a {@link Transaction} and commit it after the
+     * work completes.
      * <p>
-     * Every read observes the transaction's isolated snapshot, including its
-     * own uncommitted writes, and joins its conflict footprint. Results resolve
-     * eagerly, and every loaded {@link Record}, including the records reachable
-     * from its fields, is bound to this view, so each {@link Record#save()
-     * save} stages within the transaction. An
-     * {@link com.cinchapi.runway.access.Audience Audience} loaded through the
-     * view routes the operations it performs through the transaction as well,
-     * so access-controlled reads and writes stay within the snapshot. Writes
-     * become durable when {@link #commit()} succeeds; until then no reader
-     * outside the transaction can observe them. A {@link DeferredReference}
-     * that is first accessed within the transaction resolves within it as well.
-     * Only {@link Record Records} persisted in the database are visible;
-     * records supplied by an attached {@link AdHocDataSource} are not.
-     * </p>
-     * <p>
-     * A {@link Transaction} is confined to the thread that starts it and must
-     * be ended by exactly one of {@link #commit()} or {@link #abort()};
-     * {@link #close()} aborts whatever was not committed, so a
-     * try-with-resources block guarantees a clean end. After the transaction
-     * ends, it forwards reads and saves to the enclosing {@link Runway}, so a
-     * {@link Record} bound to it unwinds to the database scope; only another
-     * {@link #commit()} is refused.
+     * This method behaves exactly like {@link #call(Function)} for work that
+     * does not produce a result.
      * </p>
      *
-     * @author Jeff Nelson
+     * @param work the work to run
+     * @throws RetryExhaustedException if the transaction cannot commit within
+     *             the bounds of the governing {@link AtomicRetryPolicy}
      */
-    public class Transaction implements
-            PersistentDatabaseInterface,
-            AutoCloseable {
-
-        /**
-         * The connection that hosts the staged transaction and services every
-         * read, or {@code null} when this view is the {@link NullTransaction}.
-         */
-        private final Concourse concourse;
-
-        /**
-         * Whether this view owns {@link #concourse} and must release it back to
-         * the {@link #connections pool} when the transaction ends.
-         */
-        private final boolean owned;
-
-        /**
-         * The only {@link Thread} that may operate on this view.
-         */
-        private final Thread owner;
-
-        /**
-         * Whether the transaction is still active.
-         */
-        private boolean open;
-
-        /**
-         * The {@link SaveContext} for every save staged within the transaction,
-         * in staging order, so the lifecycle consequences can be dispatched at
-         * {@link #commit()} or unwound at {@link #abort()}.
-         */
-        private final List<SaveContext> saves = new ArrayList<>();
-
-        /**
-         * The {@link ConcourseProvider} that scopes a bound {@link Record
-         * Record's} operations to this transaction.
-         */
-        private final ConcourseProvider provider = new ConcourseProvider() {
-
-            @Override
-            public Concourse request() {
-                if(open) {
-                    verifyOwner();
-                    return concourse;
-                }
-                else {
-                    return connections.request();
-                }
-            }
-
-            @Override
-            public void release(Concourse connection) {
-                if(!open) {
-                    connections.release(connection);
-                }
-                else {
-                    // no-op: the Transaction owns the connection until the
-                    // transaction ends
-                }
-            }
-
-        };
-
-        /**
-         * Construct a new instance and {@link Concourse#stage() stage} a
-         * transaction on {@code concourse}.
-         *
-         * @param concourse the {@link Concourse} connection that services every
-         *            operation; must not already be in a transaction
-         * @param owned {@code true} if this view must release {@code concourse}
-         *            back to the pool when the transaction ends
-         */
-        Transaction(Concourse concourse, boolean owned) {
-            this.concourse = concourse;
-            this.owned = owned;
-            this.owner = Thread.currentThread();
-            this.open = true;
-            concourse.stage();
-        }
-
-        /**
-         * Construct the {@link NullTransaction} instance, which has no
-         * connection and stages nothing.
-         */
-        private Transaction() {
-            this.concourse = null;
-            this.owned = false;
-            this.owner = Thread.currentThread();
-            this.open = false;
-        }
-
-        @Override
-        public Selections select(Selection<?>... options) {
-            if(!open) {
-                return Runway.this.select(options);
-            }
-            verifyOwner();
-            Preconditions.checkArgument(options.length > 0);
-            DatabaseSelection<?>[] selections = Arrays.stream(options)
-                    .peek(option -> Preconditions.checkState(
-                            option.state() == Selection.State.PENDING || option
-                                    .state() == Selection.State.RESOLVED,
-                            "Selection has already been submitted"))
-                    .map(DatabaseSelection::resolve)
-                    .toArray(DatabaseSelection[]::new);
-            try (Reader reader = supportsBulkCommands
-                    ? new BatchReader(concourse)
-                    : new IncrementalReader(concourse)) {
-                for (DatabaseSelection<?> selection : selections) {
-                    if(selection.state == Selection.State.RESOLVED) {
-                        selection.setState(Selection.State.FINISHED);
-                    }
-                    else {
-                        selection.setState(Selection.State.SUBMITTED);
-                        $selectFromDatabase(reader, selection, this);
-                    }
-                }
-                reader.drain();
-            }
-            Set<Record> seen = Sets.newIdentityHashSet();
-            for (DatabaseSelection<?> selection : selections) {
-                bind(selection.get(), seen);
-            }
-            return new Selections(selections);
-        }
-
-        /**
-         * Load a record by {@code id} without knowing its class.
-         *
-         * @param id
-         * @return the loaded record
-         */
-        @Override
-        public <T extends Record> T load(long id) {
-            if(open) {
-                verifyOwner();
-                Set<Object> sections = concourse.select(Record.SECTION_KEY, id);
-                Class<T> clazz = Reflection
-                        .getClassCasted((String) Iterables.getLast(sections));
-                return load(clazz, id);
-            }
-            else {
-                return Runway.this.load(id);
-            }
-        }
-
-        /**
-         * Save all changes in the provided {@code records} within this
-         * transaction.
-         * <p>
-         * The records, and every {@link Record} linked from them, are bound to
-         * this transaction, and the staged changes become durable when
-         * {@link #commit()} succeeds. Until then, no reader outside the
-         * transaction can observe them.
-         * </p>
-         *
-         * @param records one or more {@link Record Records} to save
-         * @return {@code true} when the changes are staged
-         */
-        @Override
-        public boolean save(Record... records) {
-            return save(false, records);
-        }
-
-        /**
-         * Save all changes in the provided {@code records} within this
-         * transaction.
-         * <p>
-         * The records, and every {@link Record} linked from them, are bound to
-         * this transaction, and the staged changes become durable when
-         * {@link #commit()} succeeds. Until then, no reader outside the
-         * transaction can observe them.
-         * </p>
-         *
-         * @param preventStaleWrites if {@code true}, reject the save when any
-         *            {@link Record} in the object graph has stale data
-         * @param records one or more {@link Record Records} to save
-         * @return {@code true} when the changes are staged
-         * @throws StaleDataException if {@code preventStaleWrites} is
-         *             {@code true} and any {@link Record} has been externally
-         *             modified
-         */
-        @Override
-        public boolean save(boolean preventStaleWrites, Record... records) {
-            if(!open) {
-                return Runway.this.save(preventStaleWrites, records);
-            }
-            verifyOwner();
-            Saver saver = new IncrementalSaver(concourse);
-            SaveContext context = new SaveContext(preventStaleWrites);
-            try {
-                // NOTE: The saver is never staged or committed here because
-                // the connection is already within this transaction, whose
-                // commit is the terminal operation.
-                for (Record record : records) {
-                    Verify.that(record.overrideSave() == null,
-                            "Cannot save a Record that overrides the save"
-                                    + " pipeline within a Transaction");
-                    record.bind(this, provider);
-                    record.saveWithinTransaction(saver, context);
-                }
-                stageDeletions(saver, context);
-            }
-            catch (Throwable t) {
-                context.restore();
-                throw t;
-            }
-            saves.add(context);
-            return true;
-        }
-
-        /**
-         * Attempt to commit the transaction and make every staged write
-         * durable.
-         * <p>
-         * On success, the lifecycle consequences of each staged save (e.g.,
-         * save and delete notifications) are dispatched. On failure, every
-         * staged write is discarded; a {@link Record Record's} in-memory edits
-         * remain, the same as after a failed save. Either way, the transaction
-         * ends.
-         * </p>
-         *
-         * @return {@code true} if the transaction commits
-         */
-        public boolean commit() {
-            verify();
-            boolean committed = false;
-            try {
-                committed = concourse.commit();
-            }
-            finally {
-                end(committed);
-            }
-            return committed;
-        }
-
-        /**
-         * Abort the transaction and discard every staged write; a {@link Record
-         * Record's} in-memory edits remain, the same as after a failed save.
-         * <p>
-         * This method has no effect if the transaction already ended.
-         * </p>
-         */
-        public void abort() {
-            if(open) {
-                verifyOwner();
-                try {
-                    concourse.abort();
-                }
-                finally {
-                    end(false);
-                }
-            }
-        }
-
-        @Override
-        public void close() {
-            abort();
-        }
-
-        /**
-         * Return a private synchronous {@link Reader} for a resolver that must
-         * drive its own reads within this {@link Transaction}.
-         *
-         * @return the {@link Reader}
-         */
-        Reader syncReader() {
-            return new IncrementalReader(concourse);
-        }
-
-        /**
-         * Execute {@link Concourse#verifyOrSet(String, Object, long)
-         * verifyOrSet} within the transaction.
-         *
-         * @param key the field name
-         * @param value the value to store as the only value for {@code key} in
-         *            {@code record}
-         * @param record the record id
-         */
-        void verifyOrSet(String key, Object value, long record) {
-            concourse.verifyOrSet(key, value, record);
-        }
-
-        /**
-         * Return the {@link Runway} instance that this {@link Transaction}
-         * operates against.
-         *
-         * @return the enclosing {@link Runway}
-         */
-        Runway database() {
-            return Runway.this;
-        }
-
-        /**
-         * Bind {@code result} to this {@link Transaction}: a {@link Record} is
-         * bound along with its loaded graph, an {@link Iterable} is bound
-         * element-wise, and any other result is left alone.
-         *
-         * @param result a resolved {@link Selection} result
-         * @param seen the identity set of {@link Record Records} that are
-         *            already bound
-         */
-        private void bind(Object result, Set<Record> seen) {
-            if(result instanceof Record) {
-                ((Record) result).bindGraph(this, provider, seen);
-            }
-            else if(result instanceof Iterable) {
-                for (Object item : (Iterable<?>) result) {
-                    bind(item, seen);
-                }
-            }
-        }
-
-        /**
-         * End the transaction: dispatch or unwind the lifecycle consequences of
-         * every staged save and, if this view {@link #owned owns} the
-         * connection, release it back to the pool.
-         *
-         * @param committed {@code true} if the transaction committed
-         */
-        private void end(boolean committed) {
-            open = false;
-            if(committed) {
-                for (SaveContext context : saves) {
-                    dispatchSaveOutcomes(context);
-                }
-            }
-            else {
-                for (int i = saves.size() - 1; i >= 0; --i) {
-                    saves.get(i).restore();
-                }
-            }
-            if(owned) {
-                connections.release(concourse);
-            }
-        }
-
-        /**
-         * Verify that the transaction is still {@link #open} and that the
-         * caller is the {@link #owner} thread.
-         */
-        private void verify() {
-            verifyOwner();
-            Verify.that(open, "The Transaction has ended");
-        }
-
-        /**
-         * Verify that the caller is the {@link #owner} thread.
-         */
-        private void verifyOwner() {
-            Verify.that(Thread.currentThread() == owner,
-                    "A Transaction is confined to the thread that started it");
-        }
-
+    public void run(Consumer<Transaction> work) {
+        call(transaction -> {
+            work.accept(transaction);
+            return null;
+        });
     }
+
 }
