+++ src/main/java/com/cinchapi/runway/Transaction.java
@@ -1,0 +1,449 @@
+/*
+ * Copyright (c) 2013-2026 Cinchapi Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.cinchapi.runway;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
+
+import com.cinchapi.common.base.Verify;
+import com.cinchapi.common.reflect.Reflection;
+import com.cinchapi.concourse.Concourse;
+import com.cinchapi.runway.db.BatchReader;
+import com.cinchapi.runway.db.ConcourseProvider;
+import com.cinchapi.runway.db.IncrementalReader;
+import com.cinchapi.runway.db.IncrementalSaver;
+import com.cinchapi.runway.db.Reader;
+import com.cinchapi.runway.db.Saver;
+import com.google.common.base.Preconditions;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Sets;
+
+/**
+ * A {@link DatabaseInterface} that scopes every operation to a single ACID
+ * transaction.
+ * <p>
+ * Every read observes the transaction's isolated snapshot, including its own
+ * uncommitted writes, and joins its conflict footprint. Results resolve
+ * eagerly, and every loaded {@link Record}, including the records reachable
+ * from its fields, is bound to this view, so each {@link Record#save() save}
+ * stages within the transaction. A {@link Record}
+ * {@link #create(Class, Object...) created} through the view is bound to it as
+ * well. An {@link com.cinchapi.runway.access.Audience Audience} loaded through
+ * the view routes the operations it performs through the transaction as well,
+ * so access-controlled reads and writes stay within the snapshot. Writes become
+ * durable when {@link #commit()} succeeds; until then no reader outside the
+ * transaction can observe them. A {@link DeferredReference} that is first
+ * accessed within the transaction resolves within it as well. Only
+ * {@link Record Records} persisted in the database are visible; records
+ * supplied by an attached {@link AdHocDataSource} are not.
+ * </p>
+ * <p>
+ * A {@link Transaction} is confined to the thread that starts it and must be
+ * ended by exactly one of {@link #commit()} or {@link #abort()};
+ * {@link #close()} aborts whatever was not committed, so a try-with-resources
+ * block guarantees a clean end. After the transaction ends, it forwards reads
+ * and saves to the enclosing {@link Runway}, so a {@link Record} bound to it
+ * unwinds to the database scope; only another {@link #commit()} is refused.
+ * </p>
+ *
+ * @author Jeff Nelson
+ */
+public class Transaction implements PersistentDatabaseInterface, AutoCloseable {
+
+    /**
+     * The {@link Runway} instance that this {@link Transaction} operates
+     * against and falls through to after the transaction ends.
+     */
+    private final Runway database;
+
+    /**
+     * The connection that hosts the staged transaction and services every read,
+     * or {@code null} when this view represents the absence of a transaction.
+     */
+    private final Concourse concourse;
+
+    /**
+     * Whether this view owns {@link #concourse} and must release it back to the
+     * connection pool when the transaction ends.
+     */
+    private final boolean owned;
+
+    /**
+     * The only {@link Thread} that may operate on this view.
+     */
+    private final Thread owner;
+
+    /**
+     * Whether the transaction is still active.
+     */
+    private boolean open;
+
+    /**
+     * The {@link SaveContext} for every save staged within the transaction, in
+     * staging order, so the lifecycle consequences can be dispatched at
+     * {@link #commit()} or unwound at {@link #abort()}.
+     */
+    private final List<SaveContext> saves = new ArrayList<>();
+
+    /**
+     * The {@link ConcourseProvider} that scopes a bound {@link Record Record's}
+     * operations to this transaction.
+     */
+    private final ConcourseProvider provider = new ConcourseProvider() {
+
+        @Override
+        public Concourse request() {
+            if(open) {
+                verifyOwner();
+                return concourse;
+            }
+            else {
+                return database.provider.request();
+            }
+        }
+
+        @Override
+        public void release(Concourse connection) {
+            if(!open) {
+                database.provider.release(connection);
+            }
+            else {
+                // no-op: the Transaction owns the connection until the
+                // transaction ends
+            }
+        }
+
+    };
+
+    /**
+     * Construct a new instance and {@link Concourse#stage() stage} a
+     * transaction on {@code concourse}.
+     *
+     * @param database the {@link Runway} instance that this transaction
+     *            operates against
+     * @param concourse the {@link Concourse} connection that services every
+     *            operation; must not already be in a transaction
+     * @param owned {@code true} if this view must release {@code concourse}
+     *            back to the pool when the transaction ends
+     */
+    Transaction(Runway database, Concourse concourse, boolean owned) {
+        this.database = database;
+        this.concourse = concourse;
+        this.owned = owned;
+        this.owner = Thread.currentThread();
+        this.open = true;
+        concourse.stage();
+    }
+
+    /**
+     * Construct an instance that represents the absence of a transaction: it
+     * has no connection and stages nothing, so every operation behaves exactly
+     * as it does on {@code database}.
+     *
+     * @param database the {@link Runway} instance that this transaction
+     *            operates against
+     */
+    Transaction(Runway database) {
+        this.database = database;
+        this.concourse = null;
+        this.owned = false;
+        this.owner = Thread.currentThread();
+        this.open = false;
+    }
+
+    @Override
+    public Selections select(Selection<?>... options) {
+        if(!open) {
+            return database.select(options);
+        }
+        verifyOwner();
+        Preconditions.checkArgument(options.length > 0);
+        DatabaseSelection<?>[] selections = Arrays.stream(options)
+                .peek(option -> Preconditions.checkState(
+                        option.state() == Selection.State.PENDING
+                                || option.state() == Selection.State.RESOLVED,
+                        "Selection has already been submitted"))
+                .map(DatabaseSelection::resolve)
+                .toArray(DatabaseSelection[]::new);
+        try (Reader reader = database.supportsBulkCommands
+                ? new BatchReader(concourse)
+                : new IncrementalReader(concourse)) {
+            for (DatabaseSelection<?> selection : selections) {
+                if(selection.state == Selection.State.RESOLVED) {
+                    selection.setState(Selection.State.FINISHED);
+                }
+                else {
+                    selection.setState(Selection.State.SUBMITTED);
+                    database.$selectFromDatabase(reader, selection, this);
+                }
+            }
+            reader.drain();
+        }
+        Set<Record> seen = Sets.newIdentityHashSet();
+        for (DatabaseSelection<?> selection : selections) {
+            bind(selection.get(), seen);
+        }
+        return new Selections(selections);
+    }
+
+    /**
+     * Create a new {@link Record} of the specified {@code clazz} that is bound
+     * to this {@link Transaction}, so a direct {@link Record#save() save}
+     * stages within it.
+     * <p>
+     * The returned {@link Record} is not saved to the database until
+     * {@link Record#save()} is called. After the transaction ends, the
+     * {@link Record} operates against the enclosing {@link Runway}.
+     * </p>
+     *
+     * @param clazz the type of {@link Record} to create
+     * @param args constructor arguments for the {@link Record}
+     * @param <T> the type of {@link Record}
+     * @return the newly created {@link Record}, not yet saved
+     */
+    public <T extends Record> T create(Class<T> clazz, Object... args) {
+        if(open) {
+            verifyOwner();
+        }
+        T record = Reflection.newInstance(clazz, args);
+        record.bind(this, provider);
+        return record;
+    }
+
+    /**
+     * Load a record by {@code id} without knowing its class.
+     *
+     * @param id
+     * @return the loaded record
+     */
+    @Override
+    public <T extends Record> T load(long id) {
+        if(open) {
+            verifyOwner();
+            Set<Object> sections = concourse.select(Record.SECTION_KEY, id);
+            Class<T> clazz = Reflection
+                    .getClassCasted((String) Iterables.getLast(sections));
+            return load(clazz, id);
+        }
+        else {
+            return database.load(id);
+        }
+    }
+
+    /**
+     * Save all changes in the provided {@code records} within this transaction.
+     * <p>
+     * The records, and every {@link Record} linked from them, are bound to this
+     * transaction, and the staged changes become durable when {@link #commit()}
+     * succeeds. Until then, no reader outside the transaction can observe them.
+     * </p>
+     *
+     * @param records one or more {@link Record Records} to save
+     * @return {@code true} when the changes are staged
+     */
+    @Override
+    public boolean save(Record... records) {
+        return save(false, records);
+    }
+
+    /**
+     * Save all changes in the provided {@code records} within this transaction.
+     * <p>
+     * The records, and every {@link Record} linked from them, are bound to this
+     * transaction, and the staged changes become durable when {@link #commit()}
+     * succeeds. Until then, no reader outside the transaction can observe them.
+     * </p>
+     *
+     * @param preventStaleWrites if {@code true}, reject the save when any
+     *            {@link Record} in the object graph has stale data
+     * @param records one or more {@link Record Records} to save
+     * @return {@code true} when the changes are staged
+     * @throws StaleDataException if {@code preventStaleWrites} is {@code true}
+     *             and any {@link Record} has been externally modified
+     */
+    @Override
+    public boolean save(boolean preventStaleWrites, Record... records) {
+        if(!open) {
+            return database.save(preventStaleWrites, records);
+        }
+        verifyOwner();
+        Saver saver = new IncrementalSaver(concourse);
+        SaveContext context = new SaveContext(preventStaleWrites);
+        try {
+            // NOTE: The saver is never staged or committed here because
+            // the connection is already within this transaction, whose
+            // commit is the terminal operation.
+            for (Record record : records) {
+                Verify.that(record.overrideSave() == null,
+                        "Cannot save a Record that overrides the save"
+                                + " pipeline within a Transaction");
+                record.bind(this, provider);
+                record.saveWithinTransaction(saver, context);
+            }
+            database.stageDeletions(saver, context);
+        }
+        catch (Throwable t) {
+            context.restore();
+            throw t;
+        }
+        saves.add(context);
+        return true;
+    }
+
+    /**
+     * Attempt to commit the transaction and make every staged write durable.
+     * <p>
+     * On success, the lifecycle consequences of each staged save (e.g., save
+     * and delete notifications) are dispatched. On failure, every staged write
+     * is discarded; a {@link Record Record's} in-memory edits remain, the same
+     * as after a failed save. Either way, the transaction ends.
+     * </p>
+     *
+     * @return {@code true} if the transaction commits
+     */
+    public boolean commit() {
+        verify();
+        boolean committed = false;
+        try {
+            committed = concourse.commit();
+        }
+        finally {
+            end(committed);
+        }
+        return committed;
+    }
+
+    /**
+     * Abort the transaction and discard every staged write; a {@link Record
+     * Record's} in-memory edits remain, the same as after a failed save.
+     * <p>
+     * This method has no effect if the transaction already ended.
+     * </p>
+     */
+    public void abort() {
+        if(open) {
+            verifyOwner();
+            try {
+                concourse.abort();
+            }
+            finally {
+                end(false);
+            }
+        }
+    }
+
+    @Override
+    public void close() {
+        abort();
+    }
+
+    /**
+     * Return a private synchronous {@link Reader} for a resolver that must
+     * drive its own reads within this {@link Transaction}.
+     *
+     * @return the {@link Reader}
+     */
+    Reader syncReader() {
+        return new IncrementalReader(concourse);
+    }
+
+    /**
+     * Execute {@link Concourse#verifyOrSet(String, Object, long) verifyOrSet}
+     * within the transaction.
+     *
+     * @param key the field name
+     * @param value the value to store as the only value for {@code key} in
+     *            {@code record}
+     * @param record the record id
+     */
+    void verifyOrSet(String key, Object value, long record) {
+        concourse.verifyOrSet(key, value, record);
+    }
+
+    /**
+     * Return the {@link Runway} instance that this {@link Transaction} operates
+     * against.
+     *
+     * @return the enclosing {@link Runway}
+     */
+    Runway database() {
+        return database;
+    }
+
+    /**
+     * Bind {@code result} to this {@link Transaction}: a {@link Record} is
+     * bound along with its loaded graph, an {@link Iterable} is bound
+     * element-wise, and any other result is left alone.
+     *
+     * @param result a resolved {@link Selection} result
+     * @param seen the identity set of {@link Record Records} that are already
+     *            bound
+     */
+    private void bind(Object result, Set<Record> seen) {
+        if(result instanceof Record) {
+            ((Record) result).bindGraph(this, provider, seen);
+        }
+        else if(result instanceof Iterable) {
+            for (Object item : (Iterable<?>) result) {
+                bind(item, seen);
+            }
+        }
+    }
+
+    /**
+     * End the transaction: dispatch or unwind the lifecycle consequences of
+     * every staged save and, if this view {@link #owned owns} the connection,
+     * release it back to the pool.
+     *
+     * @param committed {@code true} if the transaction committed
+     */
+    private void end(boolean committed) {
+        open = false;
+        if(committed) {
+            for (SaveContext context : saves) {
+                database.dispatchSaveOutcomes(context);
+            }
+        }
+        else {
+            for (int i = saves.size() - 1; i >= 0; --i) {
+                saves.get(i).restore();
+            }
+        }
+        if(owned) {
+            database.provider.release(concourse);
+        }
+    }
+
+    /**
+     * Verify that the transaction is still {@link #open} and that the caller is
+     * the {@link #owner} thread.
+     */
+    private void verify() {
+        verifyOwner();
+        Verify.that(open, "The Transaction has ended");
+    }
+
+    /**
+     * Verify that the caller is the {@link #owner} thread.
+     */
+    private void verifyOwner() {
+        Verify.that(Thread.currentThread() == owner,
+                "A Transaction is confined to the thread that started it");
+    }
+
+}
